Signup a User & minor fixes in User routes
